update checkedInCout on adding member to event on dashboard page, move EventAttendance component to right place
