+++ client/src/pages/events/Events.js
@@ -6,7 +6,7 @@ import { PencilSquare, Trash, PersonAdd, ListCheck } from "react-bootstrap-icons
 
 import CreateEvent from "./components/CreateEvent";
 import AddMemberToEventModal from "./components/AddMemberToEventModal";
-import EventAttendance from "../members/EventAttendance";
+import EventAttendance from "./components/EventAttendance";
 
 import styles from "./Events.module.css";
 
@@ -22,7 +22,12 @@ const Events = () => {
   const [deleteEventId, setDeleteEventId] = useState(null);
   const [showPersonAddModal, setShowPersonAddModal] = useState([false, 0]);
   const [reqInProcess, setReqInProcess] = useState(false);
-  const [errorAlert, setErrorAlert] = useState(false);
+  const [notification, setNotification] = useState({
+    show: false,
+    color: "",
+    "message": "",
+    data: ""
+  });
 
   useEffect(() => {
     async function getEvents() {
@@ -55,7 +60,7 @@ const Events = () => {
   const create = () => {
     setFormAction("create");
     setReqInProcess(false);
-    setErrorAlert(false);
+    setNotification({ show: false, color: "", message: "" });
     setShowFormModal(true);
   };
   const createEvent = (event) => {
@@ -71,7 +76,7 @@ const Events = () => {
     };
     setSingleEvent(updatedEvent);
     setReqInProcess(false);
-    setErrorAlert(false);
+    setNotification({ show: false, color: "", message: "" });
     setShowFormModal(true);
   };
 
@@ -83,12 +88,12 @@ const Events = () => {
   const showDeleteConfirmation = (id) => {
     setDeleteEventId(id);
     setShowDeleteModal(true);
-    setErrorAlert(false);
+    setNotification({ show: false, color: "", message: "" });
   };
 
   const deleteEvent = async () => {
     setReqInProcess(true);
-    setErrorAlert(false);
+    setNotification({ show: false, color: "", message: "" });
 
     try {
       const accessToken = await getAccessTokenSilently({
@@ -114,12 +119,12 @@ const Events = () => {
       } else {
         const data = await res.json();
         console.log(data);
-        setErrorAlert(true);
+        setNotification({ show: true, color: "danger", message: "There was a problem." });
         setReqInProcess(false);
       }
     } catch (e) {
       console.log(e.message);
-      setErrorAlert(true);
+      setNotification({ show: true, color: "danger", message: "There was a problem." });
       setReqInProcess(false);
     }
   };
@@ -145,8 +150,8 @@ const Events = () => {
             setShowFormModal={setShowFormModal}
             reqInProcess={reqInProcess}
             setReqInProcess={setReqInProcess}
-            errorAlert={errorAlert}
-            setErrorAlert={setErrorAlert}
+            notification={notification}
+            setNotification={setNotification}
           />
         </Modal.Body>
       </Modal>
@@ -160,9 +165,9 @@ const Events = () => {
         </Modal.Header>
         <Modal.Body>
           Are you sure you want to delete this event?
-          {errorAlert && (
-            <Alert className="mt-3" variant="danger">
-              There was a problem. Please try again.
+          {notification.show && (
+            <Alert className="mt-3" variant={notification.color}>
+              {notification.message}
             </Alert>
           )}
         </Modal.Body>
@@ -192,8 +197,8 @@ const Events = () => {
         setShowPersonAddModal={setShowPersonAddModal}
         reqInProcess={reqInProcess}
         setReqInProcess={setReqInProcess}
-        errorAlert={errorAlert}
-        setErrorAlert={setErrorAlert}
+        notification={notification}
+        setNotification={setNotification}
       />
         <Modal show={showAttendanceModal[0]} onHide={() => setShowAttendanceModal([false, 0])}>
         <Modal.Header closeButton>
@@ -247,13 +252,13 @@ const Events = () => {
                     className={styles.icon}
                     onClick={() => {
                       setReqInProcess(false);
-                      setErrorAlert(false);
+                      setNotification({ show: false, color: "", message: "" });
                       showDeleteConfirmation(event.id);
                     }}
                   />
                   <PersonAdd className={styles.icon} onClick={() => {
                     setReqInProcess(false);
-                    setErrorAlert(false);
+                    setNotification({ show: false, color: "", message: "" });
                     setShowPersonAddModal([true, event.id]);
                   }} />
                   <ListCheck className={styles.icon} onClick={() => setShowAttendanceModal([true, event.id])} />
